Kinda working routing, except, like, not.
Firstly, because we've made osm objects into the nodes instead of nodes, we
can't usefully compute distances any more.

We'll basically have to materialise it so the vertexes are nodes once again,
but each way and relation are represented sequence of edges, n[i]->n[i+1].

This is somewhat complicated by the fact that relations can contain othe
relations, as well as nodes. Great.